Decrease updates from every second to every 20 seconds
There is currently no need for that level of precision, and it reduces
the disk writes on the home server which has been a bit too busy lately.
Perhaps a longer term solution is to see if I can introduce an SSD or
in-memory cache for InfluxDB to reduce disk writes.